Split database into separate module
