+++ cihai/core.py
@@ -7,11 +7,9 @@ import os
 
 import kaptan
 from appdirs import AppDirs
-from sqlalchemy import MetaData, create_engine
-from sqlalchemy.ext.automap import automap_base
-from sqlalchemy.orm import Session
 
-from cihai import bootstrap, exc, extend
+from cihai import exc, extend
+from cihai.db import Database
 from cihai.config import expand_config
 from cihai.constants import DEFAULT_CONFIG
 from cihai.utils import import_string, merge_dict
@@ -19,51 +17,6 @@ from cihai.utils import import_string, merge_dict
 from ._compat import string_types
 
 log = logging.getLogger(__name__)
-
-
-class Database(object):
-    def __init__(self, config):
-        self.engine = create_engine(config['database']['url'])
-
-        self.metadata = MetaData()
-        self.metadata.bind = self.engine
-        self.reflect_db()
-
-        self.session = Session(self.engine)
-
-    def reflect_db(self):
-        """
-        No-op to reflect db info.
-
-        This is available as a method so the database can be reflected
-        outside initialization (such bootstrapping unihan during CLI usage).
-        """
-        self.metadata.reflect(views=True, extend_existing=True)
-        self.base = automap_base(metadata=self.metadata)
-        self.base.prepare()
-
-    @property
-    def is_bootstrapped(self):
-        """Return True if UNIHAN and database is set up.
-
-        Returns
-        -------
-        bool :
-            True if Unihan application fixture data installed.
-        """
-        return bootstrap.is_bootstrapped(self.metadata)
-
-    #: :class:`sqlalchemy.engine.Engine` instance.
-    engine = None
-
-    #: :class:`sqlalchemy.schema.MetaData` instance.
-    metadata = None
-
-    #: :class:`sqlalchemy.orm.session.Session` instance.
-    session = None
-
-    #: :class:`sqlalchemy.ext.automap.AutomapBase` instance.
-    base = None
 
 
 class Cihai(object):
